Santander Customer Transaction Prediction
we can more accurately identify new ways to solve our most common challenge, binary classification problems such as: is a customer satisfied? Will a customer buy this product? Can a customer pay this loan? In this challenge, we need to identify which customers will make a specific transaction in the future, irrespective of the amount of money transacted.